requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
openai==0.28.1
plotly==5.17.0
pypdf==3.17.0
python-docx==0.8.11
pydantic==2.4.2
python-dotenv==1.0.0
altair==5.1.1
watchdog==3.0.0
requests==2.31.0
pytest==7.4.0
pytest-html==3.2.0
